+++ backend/src/lib/bounded-contexts/marketing/marketing/tests/__tests__/change-user-email/change-user-email.steps.ts
@@ -7,9 +7,7 @@ import { ChangeUserEmailCommand } from '@src/lib/bounded-contexts/marketing/mark
 import { ChangeUserEmailCommandHandler } from '@src/lib/bounded-contexts/marketing/marketing/application/command-handlers/change-user-email.command-handler';
 import { MockUserWriteRepo } from './change-user-email-write-repo.mock';
 import { mockAsyncLocalStorageGet } from '../../mocks/mockAsynLocalStorageGet.mock';
-import { UserEntityBuilder } from '../../builders/user-entity.builder';
 import { UserEntity } from '../../../domain/user.entity';
-import { UserPropsBuilder } from '../../builders/user-props.builder';
 
 describe('Change user email feature test', () => {
   it('Changed user email successfully,', async () => {
@@ -29,12 +27,6 @@ describe('Change user email feature test', () => {
     const result = await updateUserEmailHandler.execute(updateUserEmailCommand);
 
     //then
-    const userProps = new UserPropsBuilder()
-      .withId(userId)
-      .withEmail(email)
-      .withCompletedTodos(completedTodos)
-      .build();
-
     expect(mockUpdateUserWriteRepo.mockGetByIdMethod).toHaveBeenCalledWith(
       new Domain.UUIDv4(userId),
     );
@@ -44,12 +36,16 @@ describe('Change user email feature test', () => {
 
     const userAggregate =
       mockUpdateUserWriteRepo.mockUpdateMethod.mock.calls[0][0];
-    expect(userAggregate.props).toEqual(userProps);
+    expect(userAggregate.toPrimitives()).toEqual({
+      id: userId,
+      email,
+      completedTodos,
+    });
     expect(result.value).toBe(undefined);
   });
 
   it('Changed user email failed, repo error', async () => {
-    const { email, userId, completedTodos } = UPDATE_USER_REPO_ERROR_CASE;
+    const { email, userId } = UPDATE_USER_REPO_ERROR_CASE;
     // given
     const mockUpdateUserWriteRepo = new MockUserWriteRepo();
     mockAsyncLocalStorageGet(userId);