+++ backend/src/lib/bounded-contexts/todo/todo/domain/user-id.value-object.ts
@@ -1,19 +1,19 @@
 import { Domain, Either, ok } from 'ddd-tactical-core-boilerplate';
 
 interface UserIdProps {
-  id: Domain.UUIDv4;
+  id: string;
 }
 
 export class UserIdVO extends Domain.ValueObject<UserIdProps> {
   get id(): Domain.UUIDv4 {
-    return this.props.id;
+    return Domain.UUIDv4.fromString(this.props.id);
   }
 
   private constructor(props: UserIdProps) {
     super(props);
   }
 
-  public static create(props: UserIdProps): Either<UserIdVO, never> {
-    return ok(new UserIdVO(props));
+  public static create(props: { id: Domain.UUIDv4 }): Either<UserIdVO, never> {
+    return ok(new UserIdVO({ id: props.id.toString() }));
   }
 }